타입이 정해지지 않은 위젯은 post 대상에서 제외함
